chore: close "develop" branch
As discussed in #47, we are moving to a "master"-based development
workflow and retiring the "develop" branch. This commit cleans out the
files and adds a pointer to the README telling people to consult master
instead; in this way, it will be hard for anybody to accidentally target
the "develop" branch because they'll almost certainly run into
conflicts.

Closes: https://github.com/liferay/liferay-npm-tools/issues/47