chore: move check-source-formatting files into packages/ subdirectory
In service of https://github.com/liferay/liferay-npm-tools/issues/98 we
are going to start using the check-source-formatting logic on ".jsp" and
possibly ".soy" files too. Because I don't want to break the blame
information, I am going to import the history into the liferay-npm-tools
repo and store the files under "packages/check-source-formatting/"; then
I'll proceed to extract the functionality needed into
liferay-npm-scripts where we can modify and adapt it without actually
maintaining and releasing a separate package.

This commit, then, is the result of:

    git remote add csf https://github.com/liferay/liferay-frontend-source-formatter
    git fetch csf
    git checkout 84d6aa5f836c28ca9517870e8bc407e26dae7e81
    git clean -xdf
    mkdir -p packages/check-source-formatting
    git mv LICENSE \
        .gitignore \
        .npmignore \
        .travis.yml \
        README.md \
        appveyor.yml \
        bin \
        gulpfile.js \
        lib \
        package-lock.json \
        package.json \
        test \
        packages/check-source-formatting