fix: make <portlet:namespace /> substitution match length
We need to preserve width to stop Prettier from wrapping in the wrong
place, so substitute like this:

    <portlet:namespace/>
    _PORTLET_NAMESPACE__

    <portlet:namespace />
    _PORTLET_NAMESPACE___

    <portlet:namespace   />
    _PORTLET_NAMESPACE_____

    etc

In practice, we probably won't see all those variations because I bet
there is SF logic somewhere that is forcing them to always be of the
second form, but it seems harmless enough.

Also in practice, we might just want to set Prettier to have an
(effectively) infinite wrap width, so that it never breaks things across
lines, at which point we can remove all this.