feat: add restoreTags()
Gives us a way to round-trip:

  source
    → substituteTags()
      → formatted source
        → restoreTags()
          → formatted source with original tags back in place

Note that in implementing this I realized that we needed a way to
distinguish fake placeholder comments from real ones, and the obvious
trick of using the existing ID_START and ID_END codepoints doesn't work
because it breaks the "shape" of the placeholder; eg. given:

    <%\n
    %>

if we naively transform that to:

    /*{ID_START}\n
    {ID_END}*/

we've now made the first and last line one column wider. We could try to
adjust it but things get messy quickly. It's easier to just filler the
whole thing with a unique filler character, which is why
`toWhitespace()` is now `toFiller()`.