feat: don't create temp files in src directory
A new target has been added to portal's gradle build named fastDeploy.
That target leverages gradle in watch mode, which causes the build to
retrigger whenever a source file is changed.

As liferay-npm-scripts modifies .babelrc, .npmbundlerrc and similar
files directly in place (in the scr) gradle watch is fooled into
thinking that source files have changed, thus rerunning the build
multiple times.

With this commit we avoid modifying file in the source folder and create
modified versions of them in a temporary folder, pointing the tools
there when they are executed. This solves the problem.