feat(scripts)!: check desired filename format
This is just a proof-of-concept at this point because we haven't
actually decided on the rules that we want to enforce.

Note that I am just doing this for SCSS files at this point because we
might even want to pursue the alternative of baking this directly into
our ESLint rules for our JS files.

FWIW, a quick survey of existing names in liferay-portal shows:

- 68 files using uppercase (ie. CamelCase) names:
  https://gist.github.com/wincent/6d62f155c01dee91c9dd5f0f6a533d8c
- 104 files using kebab-case names:
  https://gist.github.com/wincent/c7d39715cba6f794f558b71538bdc12a
- 758 files using neither of the above (generally, that means 0 or more
  underscores):
  https://gist.github.com/wincent/df6e2f276d91bb34a89cb17c7cd8ee99

Test plan: run in liferay-portal; each `getPaths` call takes about 4
seconds on my machine even though that means scanning the entire repo --
we could memoize that or make it "smarter" in CI (ie. by using the
changed paths detection to short-circuit traversal instead of doing it
at the end as a filtering step).