fix(scripts): attempt to fix Git invocation on Windows
Troubleshooting in the dark here. Error log:

https://travis-ci.org/github/liferay/liferay-npm-tools/jobs/672201279

Shows failure here:

    run(): command `git commit -m a -- a modules` exited with status 128.

But notably, the preceding `git init` and `git add a modules` both seem
to work. This makes me wonder if `--` is processed differently on
Windows. Let's try without it.

If this works, there is one more place in the codebase where we use "--"
(in `filterChangedFiles()` itself), so we'll need to find a way to do
the same without using "--"; might be tricky because we definitely need
that, to handle repos/branches which might not have "modules/private").